Energy System in progress
Energy Voxels (White) created
Changed Voxels' materials. Now they have a subtle transparency effect.
Fade out added when Game Over or Game Win.
Player now travels through USB cables.
Electric FX added when moving through USB Cables
Added DVI steel base.
Added shield protection system when player received an attack.
Removed VSync in quality settings

+++ Assets/Scripts/GamePlay/AI/SpiderAI/SpiderAIBehaviour.cs
@@ -13,7 +13,7 @@ public class SpiderAIBehaviour : EnemyBaseAIBehaviour
     private SpiderBlackboard spiderBlackboard;
 
     [Header("Spider Settings")]
-    public int biteDamage = 10;
+    public float biteDamage = 10f;
     public float playerDetectionDistance = 5f;
     public int spidersAttackingThreshold = 3;
     public float wrongColorDamageModifier = 0.5f;
