+++ Assets/Scripts/GamePlay/Player/States/PlayerSpecialState.cs
@@ -15,7 +15,7 @@ public class PlayerSpecialState : PlayerBaseState {
         return blackboard.idleState;
     }
 
-    public override PlayerBaseState TakeDamage(int damage, bool triggerDamageAnim = true, bool whiteBlink = true)
+    public override PlayerBaseState TakeDamage(float damage, bool triggerDamageAnim = true, bool whiteBlink = true)
     {
         //can not take damage during this state
         return null;
